Refactor unit & integration tests
