+++ src/components/utils/__integrations__/onfidoApi/video.integration.js
@@ -1,15 +1,15 @@
 import fs from 'fs'
-import { requestChallenges, uploadLiveVideo } from '../onfidoApi'
+import { requestChallenges, uploadLiveVideo } from '../../onfidoApi'
 import {
   getTestJwtToken,
   checkForExpectedFileUploadProperties,
   COMMON_FILE_UPLOAD_PROPERTIES,
-} from './helpers'
-import { API_URL, PATH_TO_RESOURCE_FILES } from './testUrls'
+} from '../helpers'
+import { API_URL, PATH_TO_RESOURCE_FILES } from '../helpers/testUrls'
 import {
   EXPIRED_JWT_TOKEN,
   EXPECTED_EXPIRED_TOKEN_ERROR,
-} from './mockExpiredJwt'
+} from '../helpers/mockExpiredJwt'
 
 let jwtToken = null
