+++ src/components/utils/__tests__/string.test.js
